Added `convert` method to AbstractModel class.
I noticed that one is able to send a `convert` action to convert a Quote to an Invoice. However, I could not find a way to call it from the SDK, and so I added the function. Alternatively, I would add something like this:

public function action($action)
{
    return $this->sendAction($action);
}

Where `action` could be anything.